- first draft

+++ Hackathon/Services/IssuesService.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using Hackathon.Abstractions.Repositories;
 using Hackathon.Abstractions.Services;
+using Hackathon.Common;
 using Hackathon.Models;
 using Hackathon.Models.DTOs;
 using System.Collections.Generic;
@@ -23,7 +24,7 @@ namespace Hackathon.Services
             _photoRepository = photoRepository;
         }
 
-        public async Task<List<IssuesDto>> GetAllIssuesAsync()
+        public async Task<Result<List<IssuesDto>>> GetAllIssuesAsync()
         {
             var issuesDtoList = new List<IssuesDto>();
 
@@ -42,28 +43,39 @@ namespace Hackathon.Services
                 });
             }
 
-            return issuesDtoList;
+            return Result.OK(issuesDtoList);
         }
 
-        public async Task<IssuesDto> GetIssuesByIdAsync(int id)
+        public async Task<Result<IssuesDto>> GetIssuesByIdAsync(int id)
         {
             var issue = await _issuesRepository.GetIssuesByIdAsync(id);
+
+            if(issue == null)
+            {
+                return Result.NotFound<IssuesDto>();
+            }
+
             var photo = await _photoRepository.GetPhotoAsync(issue.PhotoName);
 
-            return new IssuesDto
+            return Result.OK(new IssuesDto
             {
                 Id = issue.Id,
                 Details = issue.Details,
                 PinId = issue.PinId,
                 Photo = photo.PhotoByteArray
-            };
+            });
         }
 
-        public async Task<List<IssuesDto>> GetIssuesByPinIdAsync(int pinId)
+        public async Task<Result<List<IssuesDto>>> GetIssuesByPinIdAsync(int pinId)
         {
             var issuesDtoList = new List<IssuesDto>();
 
             var issuesList = await _issuesRepository.GetIssuesByPinIdAsync(pinId);
+
+            if(issuesList == null)
+            {
+                return Result.BadRequest<List<IssuesDto>>().AddErrors("pinId", "The pinId is invalid");
+            }
 
             foreach (var issue in issuesList)
             {
@@ -77,10 +89,10 @@ namespace Hackathon.Services
                     });             
             }
 
-            return issuesDtoList;
+            return Result.OK(issuesDtoList);
         }
 
-        public async Task<IssuesDto> CreateIssuesAsync(IssuesDto issuesDto)
+        public async Task<Result<IssuesDto>> CreateIssuesAsync(IssuesDto issuesDto)
         {
             var photo = await _photoRepository.SavePhotoAsync(issuesDto.Photo);
 
@@ -95,24 +107,28 @@ namespace Hackathon.Services
 
             if (await _issuesRepository.CreateIssuesAsync(issueEntity))
             {
-                return new IssuesDto 
+                return Result.OK(new IssuesDto 
                 {
                     Id = issueEntity.Id,
                     Photo = photo.PhotoByteArray,
                     PinId = issueEntity.PinId,
                     Details = issueEntity.Details
-                };
+                });
             }
-            return default(IssuesDto);
+            return Result.InternalServerError(default(IssuesDto));
         }
 
-        public async Task<IssuesDto> UpdateIssuesAsync(int id, IssuesDto issuesDto)
+        public async Task<Result<IssuesDto>> UpdateIssuesAsync(int id, IssuesDto issuesDto)
         {
             var issueEntity = await _issuesRepository.GetIssuesByIdAsync(id);
+            if(issueEntity == null)
+            {
+                return Result.BadRequest<IssuesDto>().AddErrors("id", "The id is not valid");
+            }
 
             if (!await _photoRepository.UpdatePhotoAsync(issueEntity.PhotoName, issuesDto.Photo))
             {
-                return default(IssuesDto);
+                return Result.InternalServerError(default(IssuesDto));
             }
 
             issueEntity.PinId = issuesDto.PinId;
@@ -122,31 +138,35 @@ namespace Hackathon.Services
 
             if (await _issuesRepository.UpdateIssuesAsync(issueEntity))
             {
-                return new IssuesDto
+                return Result.OK(new IssuesDto
                 {
                     Id = issueEntity.Id,
                     Details = issueEntity.Details,
                     PinId = issueEntity.PinId,
                     Photo = photo.PhotoByteArray
-                };
+                });
             }
-            return default(IssuesDto);
+            return Result.InternalServerError(default(IssuesDto));
         }
 
-        public async Task<bool> DeleteIssuesAsync(int id)
+        public async Task<Result<bool>> DeleteIssuesAsync(int id)
         {
             var issue = await _issuesRepository.GetIssuesByIdAsync(id);
+            if(issue == null)
+            {
+                return Result.NotFound<bool>().AddErrors("id","The id is invalid");
+            }
 
             if(!await _photoRepository.DeletePhotoAsync(issue.PhotoName))
             {
-                return false;
+                return Result.InternalServerError(false);
             }
 
             if (await _issuesRepository.DeleteIssuesAsync(id))
             {
-                return true;
+                return Result.OK(true);
             }
-            return false;
+            return Result.InternalServerError(false);
         }
     }
 }
